+++ cmd/root_test.go
@@ -384,6 +384,13 @@ func TestBadScheme(t *testing.T) {
 	assert.Contains(t, err, "unknown database scheme")
 }
 
+func TestPattern(t *testing.T) {
+	output := captureOutput(func() { runCmd([]string{fileUrl("min-count.txt"), "--pattern", `\stest[12]`, "--show-data"}) })
+	assert.NotContains(t, output, "test1")
+	assert.Contains(t, output, "test2")
+	assert.NotContains(t, output, "test3")
+}
+
 func TestShowData(t *testing.T) {
 	output := captureOutput(func() { runCmd([]string{fileUrl("email.txt"), "--show-data"}) })
 	assert.Contains(t, output, "test@example.org")