+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Added `--only` and `--except` options
 - Added experimental `--min-count` option
+- Added experimental `--pattern` option
 - Fixed line counts
 
 ## 0.1.5 (2022-09-12)